[XLA] Move shape_with_output_layout and seed fields into ExecutionOptions.
This simplifies the execution API by getting rid of two default params.

Also change HloModuleConfig so it stores each of the fields of
ExecutionOptions individually, instead of keeping an instance of the
ExecutionOptions proto.

This is necessary because HloModuleConfig already has a field derived
from shape_with_output_layout -- if we stored the ExecutionOptions proto
in HloModuleConfig, its shape wouldn't necessarily match the shape we
already have.
Change: 146477669