Use tf. for all symbols.
Change base_dir to exclude tensorflow/.

Improve the doc generator's handling of defaults in argument lists. Use an index to look up the preferred name for objects that are passed around as defaults which has a canonical TensorFlow name. Use ast and codegen to reproduce what was in the actual code for everything else.

Also simplify @{} replace logic.
Change: 146750326