[XLA] Read GPU architecture from StreamExecutor, rather than a flag.
Previously, we read the GPU architecture from the --gpu_architecture
flag, which defaulted to compute_35.  We'd then use this value when
choosing a libdevice file and when telling LLVM which GPU architecture
we're compiling for.

Now we read this value from the StreamExecutor for the device on which
we're going to run our computation.

This change also adds more supported GPU architectures to the GPU
backend, so we choose the right libdevice for your GPU.

This change is necessary before we can begin emitting fp16 arithmetic
and other sm_60+ ops.
Change: 147971326