Undo a breaking change to the public API; while retaining the bugfix of #8504.
As it turns out, the intent of constructing tf.layers._Layer with reuse=True
is to reuse the subscope containing the layer's name.  This was broken before the recent change fixing #8504, but that change didn't actually fix the problem (it introduced a different bug!).  Here we finish fixing #8504.

The following will now work correctly:

a = Dense(3, name="a")
a.apply(...)  # scope of a is "a"

a_reuse = Dense(3, name="a", _reuse=True)
a_reuse.apply(...)  # scope of a_reuse is "a"

the errant behavior was that a_reuse scope became just "".

The alternative behavior can still be recreated via:

a_reuse = Dense(..., _reuse=True)
a_reuse.apply(..., scope=tf.get_variable_scope())  # scope of a_reuse is "".
Change: 150947095